chore: refactor ggml.h
